improvement: add dry-run option. (#3)
* improvement: add dry-run option.

* remove echo find. Add debug context.

* simplify help.

+++ exifsort.sh
@@ -41,6 +41,10 @@ FILETYPES=("*.jpg" "*.jpeg" "*.png" "*.tif" "*.tiff" "*.gif" "*.xcf" "*.mp4" "*.
 #
 DIR_BLACKLIST=('*lost*' '*noexif*' '*duplicates*' '*slideshows*' '*raw*')
 #
+# Make no changes unless DRY_RUN is set to false.
+#
+DRY_RUN="true"
+#
 # Optional: Prefix of new top-level directory to kjmove sorted photos to.
 # if you use MOVETO, it MUST have a trailing slash! Can be a relative pathspec, but an
 # absolute pathspec is recommended.
@@ -123,8 +127,10 @@ if [[ "$1" == "doAction" && "$2" != "" ]]; then
             echo " Using LMDATE: $DATETIME"
         else
             echo " Moving to ./noexif/"
-            echo mkdir -p "${MOVETO}noexif" && echo mv -n "$2" "${MOVETO}noexif/"
-            mkdir -p "${MOVETO}noexif" && mv -n "$2" "${MOVETO}noexif/"
+            echo DEBUG: mkdir -p "${MOVETO}noexif" && echo mv -n "$2" "${MOVETO}noexif/"
+            if [[ "$DRY_RUN" == "false" ]]; then
+                mkdir -p "${MOVETO}noexif" && mv -n "$2" "${MOVETO}noexif/"
+            fi
             exit
         fi
     else
@@ -186,8 +192,10 @@ if [[ "$1" == "doAction" && "$2" != "" ]]; then
     # Fix permissions
     chmod 644 "$2"
 
-    echo mkdir -p "${MOVETO}${DIRNAME}" && echo mv -n "$2" "${MOVETO}${DIRNAME}${MVCMD}"
-    mkdir -p "${MOVETO}${DIRNAME}" && mv -n "$2" "${MOVETO}${DIRNAME}${MVCMD}"
+    echo DEBUG: mkdir -p "${MOVETO}${DIRNAME}" && echo mv -n "$2" "${MOVETO}${DIRNAME}${MVCMD}"
+    if [[ "$DRY_RUN" == "false" ]]; then
+        mkdir -p "${MOVETO}${DIRNAME}" && mv -n "$2" "${MOVETO}${DIRNAME}${MVCMD}"
+    fi
     echo -e "INFO: done.\n"
     exit
 fi;
@@ -226,7 +234,7 @@ for x in "${FILETYPES[@]}"; do
     fi
 
     # Run
-    echo find . \( -regextype posix-awk -regex "./[0-9]{4}" $dir_blacklist \) -o -type f -iname "$x" -print0 -exec sh -c "$0 doAction \"{}\"" \;
+    find . \( -regextype posix-awk -regex "./[0-9]{4}" $dir_blacklist \) -o -type f -iname "$x" -print0 -exec sh -c "$0 doAction \"{}\"" \;
     echo "... end of $x"
 done
 
@@ -237,7 +245,8 @@ find . -iname '.thm' -delete
 echo "INFO: Removing Thumbs.db files ... "
 find . -name Thumbs.db -delete
 echo "INFO: done."
-echo "INFO: Cleaning up empty directories ... " find . -empty -delete
+echo "INFO: Cleaning up empty directories ... "
+find . -empty -delete
 echo "INFO: done."
 echo "INFO: Recreating dump directory ... "
 mkdir dump
